+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.2.12 - ?
 ===================
 - feat: Show selected service and method in TUI
+- feat: Show relative line numbers in editor
 
 0.2.11 - 27 Dec 2025
 ===================